refactor: изменить структуру директории
