+++ src/Service/ResoursesService/ResourcesService.ts
@@ -1,54 +1,4 @@
-import axios from 'axios';
-
-import api from '../Axios';
-import { AuthResponse } from './MainResponse';
-
-const API_URL = 'http://localhost:3000';
-
-export default class MainApi {
-  _checkResponse(res) {
-    if (res.ok) {
-      return res.json();
-    }
-    return Promise.reject(`Ошибка: ${res.status}`);
-  }
-
-  static async registration(email: string, password: string): Promise<axios.AxiosResponse<AuthResponse>> {
-    return api.post<AuthResponse>('/signup', {email, password});
-
-    // return fetch(`${this._URL}signup`, {
-    //   method: 'POST',
-    //   credentials: 'include',
-    //   headers: {
-    //     'Content-Type': 'application/json'
-    //   },
-    //   body: JSON.stringify({password, email})
-    // })
-    // .then(this._checkResponse)
-  }
-
-  static async login(email: string, password: string): Promise<axios.AxiosResponse<AuthResponse>> {
-    return api.post<AuthResponse>('/signin', {email, password});
-    
-    // fetch(`${this._URL}signin`, {
-    //   method: 'POST',
-    //   credentials: 'include',
-    //   headers: {
-    //     'Content-Type': 'application/json'
-    //   },
-    //   body: JSON.stringify({password, email})
-    // })
-    // .then(this._checkResponse)
-  }
-
-  static async logout(): Promise<void> {
-    return api.post('/signout')
-  }
-
-  static async checkAuth(): Promise<axios.AxiosResponse<AuthResponse>> {
-    return axios.get<AuthResponse>(`${API_URL}/refresh`, { withCredentials: true })
-  }
-
+export default class ResoursesService {
   getContent(token) {
     return fetch(`${this._URL}users/me`, {
       credentials: 'include',
